+++ packages/process/lib/Polysemy/Process/Effect/ProcessOutput.hs
@@ -1,0 +1,6 @@
+module Polysemy.Process.Effect.ProcessOutput where
+
+data ProcessOutput a :: Effect where
+  Chunk :: ByteString -> ByteString -> ProcessOutput a m ([a], ByteString)
+
+makeSem ''ProcessOutput