+++ packages/process/test/Polysemy/Process/Test/ProcessTest.hs
@@ -2,7 +2,6 @@
 
 module Polysemy.Process.Test.ProcessTest where
 
-import qualified Data.ByteString as ByteString
 import Polysemy.Async (asyncToIOFinal)
 import Polysemy.Conc.Interpreter.Race (interpretRace)
 import Polysemy.Resume (resuming)
@@ -12,21 +11,34 @@ import System.Process.Typed (ProcessConfig)
 
 import qualified Polysemy.Process.Effect.Process as Process
 import Polysemy.Process.Effect.Process (withProcess)
-import Polysemy.Process.Interpreter.ProcessIOE (interpretProcessIOE)
+import Polysemy.Process.Interpreter.ProcessStd (interpretProcessByteString, interpretProcessTextLines)
 
 config :: ProcessConfig () () ()
 config =
   Process.proc "cat" []
 
+messageLines :: [Text]
+messageLines =
+  replicate 4 "line"
+
 message :: ByteString
 message =
-  ByteString.replicate 10 120
+  encodeUtf8 (unlines messageLines)
 
 test_process :: UnitTest
 test_process =
-  runTestAuto $ interpretRace $ asyncToIOFinal $ interpretProcessIOE True 10 config do
+  runTestAuto $ interpretRace $ asyncToIOFinal $ interpretProcessByteString True 10 config do
     withProcess do
       response <- resuming (pure . show) do
         Process.send message
         Process.recv
       message === response
+
+test_processLines :: UnitTest
+test_processLines =
+  runTestAuto $ interpretRace $ asyncToIOFinal $ interpretProcessTextLines True 10 config do
+    withProcess do
+      response <- resuming (pure . pure . show) do
+        Process.send message
+        replicateM 4 Process.recv
+      messageLines === response